fix: Change location of config files

+++ scripts/update.sh
@@ -1,24 +1,17 @@
 #!/bin/bash
 
 docker pull l1ving/heartbeat:latest
-CONTAINER_ID="$(docker ps -f name=heartbeat --format "{{.ID}}" | head -n 1)"
 
-echo "Stopping container $CONTAINER_ID"
-docker stop "$CONTAINER_ID"
-docker rm "$CONTAINER_ID"
+if [[ "$1" != "FIRST_RUN" ]]; then
+  echo "MAKE SURE TO CREATE 'token' inside \''$HOME'/heartbeat\'"
+  CONTAINER_ID="$(docker ps -f name=heartbeat --format "{{.ID}}" | head -n 1)"
+  echo "Stopping container $CONTAINER_ID"
+  docker stop "$CONTAINER_ID"
+  docker rm "$CONTAINER_ID"
+fi
 
 docker run --name heartbeat \
   -e ADDRESS="localhost:6011" \
-  --mount type=bind,source="$HOME"/heartbeat,target=/heartbeat-files \
+  --mount type=bind,source="$HOME"/heartbeat,target=/heartbeat/config \
   --network host -d \
   l1ving/heartbeat
-
-cd "$HOME" || {
-  echo "cd failed"
-  exit 1
-}
-
-rm -rf heartbeat-tmp
-git clone --depth 1 https://github.com/l1ving/heartbeat heartbeat-tmp
-rm -rf heartbeat/www
-cp -r heartbeat-tmp/www heartbeat/www
